pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "trading-bot-rag"
version = "0.1.0"
description = "News-driven trading bot with RAG and conflict resolution (Steps 1-5 MVP)"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "AI Trading Team", email = "team@example.com"}
]

dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-multipart==0.0.6",
    "numpy==1.24.3",
    "faiss-cpu==1.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "httpx==0.25.1",
]

[project.urls]
Homepage = "https://github.com/yourusername/trading-bot-rag"
Repository = "https://github.com/yourusername/trading-bot-rag.git"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
